+++ course_attendance_service/course_attendance_service/workflows/external/zoom_api/utils.py
@@ -1,0 +1,6 @@
+import urllib.parse
+
+
+def double_encoder(uuid):
+    uuid_single_encode = urllib.parse.quote_plus(uuid)
+    return (urllib.parse.quote_plus(uuid_single_encode))
